Lab update.

+++ src/scripts/copy-code.sh
@@ -4,8 +4,11 @@ cp -r ~/Desktop/src /tmp
 while read p <&3; do
 	set $p
 	echo $2
-	sshpass -pmonkey_pass ssh -t -oStrictHostKeyChecking=no monkey_user@$2 'echo monkey_pass | sudo -S rm -rf ~/Desktop/src; exit'
-	echo "copying"
-	sshpass -pmonkey_pass scp -oStrictHostKeyChecking=no -r /tmp/src monkey_user@$2:~/Desktop/
-	echo "finished"
+	if [ "$2" != "b130" ] 
+	then
+		sshpass -pmonkey_pass ssh -t -oStrictHostKeyChecking=no monkey_user@$2 'echo monkey_pass | sudo -S rm -rf ~/Desktop/src; exit'
+		echo "copying"
+		sshpass -pmonkey_pass scp -oStrictHostKeyChecking=no -r /tmp/src monkey_user@$2:~/Desktop/
+		echo "finished"
+	fi
 done 3< /etc/hosts
